[Security Solution] Fix: host popover issue 187338 (#187848)
## Summary

When clicking on the "+6 more" icon in the "Last Command" column, a
popover appears showing six items, which is as expected. However, there
is also a text label stating "1 more not shown." This seems confusing
because all six items are already displayed.

Solution: Remove the `"n more not shown."` label at the end of the list
rendered in the popover component. For cases where there are a lot of
items to render, set a max height value for the list so it becomes
scrollable.

<img width="410" alt="Screenshot 2024-07-02 at 12 08 20"
src="https://github.com/elastic/kibana/assets/61565784/c4c29607-ba7d-4060-9d0a-4531d9e48614">

<img width="284" alt="Screenshot 2024-07-09 at 12 42 31"
src="https://github.com/elastic/kibana/assets/61565784/c0a2c1e1-b44b-46b9-a84f-173213f267cd">

### Checklist

Delete any items that are not applicable to this PR.

- [ ] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

(cherry picked from commit 3ac173e140083b61b8c6caa126d5a31e23122c6e)
