+++ x-pack/plugins/security_solution/public/common/components/tables/helpers.tsx
@@ -4,15 +4,24 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
 import { FormattedMessage } from '@kbn/i18n-react';
-import { EuiLink, EuiPopover, EuiToolTip, EuiText, EuiTextColor } from '@elastic/eui';
+import { euiThemeVars } from '@kbn/ui-theme';
+import {
+  EuiLink,
+  EuiPopover,
+  EuiToolTip,
+  EuiText,
+  EuiTextColor,
+  EuiFlexGroup,
+  EuiFlexItem,
+} from '@elastic/eui';
 import styled from 'styled-components';
 import { SecurityCellActions, CellActionsMode, SecurityCellActionsTrigger } from '../cell_actions';
 import { escapeDataProviderId } from '../drag_and_drop/helpers';
 import { defaultToEmptyTag, getEmptyTagValue } from '../empty_value';
 import { MoreRowItems } from '../page';
-import { MoreContainer } from '../../../timelines/components/field_renderers/field_renderers';
+import { MoreContainer } from '../../../timelines/components/field_renderers/more_container';
 
 const Subtext = styled.div`
   font-size: ${(props) => props.theme.eui.euiFontSizeXS};
@@ -62,8 +71,8 @@ export const getRowItemsWithActions = ({
           fieldName={fieldName}
           values={values}
           idPrefix={idPrefix}
-          maxOverflowItems={maxOverflow}
           overflowIndexStart={displayCount}
+          maxOverflowItems={maxOverflow}
         />
       </>
     ) : (
@@ -78,17 +87,21 @@ interface RowItemOverflowProps {
   fieldName: string;
   values: string[];
   idPrefix: string;
-  maxOverflowItems: number;
   overflowIndexStart: number;
+  maxOverflowItems: number;
 }
 
 export const RowItemOverflowComponent: React.FC<RowItemOverflowProps> = ({
   fieldName,
   values,
   idPrefix,
-  maxOverflowItems = 5,
   overflowIndexStart = 5,
+  maxOverflowItems = 5,
 }) => {
+  const maxVisibleValues = useMemo(
+    () => values.slice(0, maxOverflowItems + 1),
+    [values, maxOverflowItems]
+  );
   return (
     <>
       {values.length > overflowIndexStart && (
@@ -97,23 +110,30 @@ export const RowItemOverflowComponent: React.FC<RowItemOverflowProps> = ({
             <MoreContainer
               fieldName={fieldName}
               idPrefix={idPrefix}
-              values={values}
+              values={maxVisibleValues}
               overflowIndexStart={overflowIndexStart}
               moreMaxHeight="none"
             />
-
-            {values.length > overflowIndexStart + maxOverflowItems && (
-              <p data-test-subj="popover-additional-overflow">
-                <EuiTextColor color="subdued">
-                  {values.length - overflowIndexStart - maxOverflowItems}{' '}
-                  <FormattedMessage
-                    id="xpack.securitySolution.tables.rowItemHelper.moreDescription"
-                    defaultMessage="more not shown"
-                  />
-                </EuiTextColor>
-              </p>
-            )}
           </EuiText>
+          {values.length > overflowIndexStart + maxOverflowItems && (
+            <EuiFlexGroup
+              css={{ paddingTop: euiThemeVars.euiSizeM }}
+              data-test-subj="popover-additional-overflow"
+            >
+              <EuiFlexItem>
+                <EuiText size="xs">
+                  <EuiTextColor color="subdued">
+                    {values.length - overflowIndexStart - maxOverflowItems}{' '}
+                    <FormattedMessage
+                      data-test-subj="popover-additional-overflow-text"
+                      id="xpack.securitySolution.tables.rowItemHelper.moreDescription"
+                      defaultMessage="more not shown"
+                    />
+                  </EuiTextColor>
+                </EuiText>
+              </EuiFlexItem>
+            </EuiFlexGroup>
+          )}
         </Popover>
       )}
     </>
