+++ x-pack/plugins/security_solution/public/timelines/components/side_panel/new_user_detail/columns.tsx
@@ -9,7 +9,7 @@ import { css } from '@emotion/react';
 import React from 'react';
 import { euiThemeVars } from '@kbn/ui-theme';
 import type { EuiBasicTableColumn } from '@elastic/eui';
-import { DefaultFieldRenderer } from '../../field_renderers/field_renderers';
+import { DefaultFieldRenderer } from '../../field_renderers/default_renderer';
 import type { ManagedUsersTableColumns, ManagedUserTable } from './types';
 import * as i18n from './translations';
 import { defaultToEmptyTag } from '../../../../common/components/empty_value';
